backend/requirements.txt
aiofiles==0.7.0
alembic==1.5.8
asyncpg==0.22.0
attrs==20.3.0
autopep8==1.5.7
bcrypt==3.2.0
billiard==3.6.4.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.1
fastapi==0.63.0
flake8==3.9.1
greenlet==1.1.0
h11==0.12.0
httpcore==0.13.2
idna==2.10
iniconfig==1.1.1
kombu==5.0.2
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
passlib==1.7.4
Pillow==8.2.0
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.1
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
requests==2.25.1
rfc3986==1.4.0
rsa==4.7.2
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.4.14
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
uvicorn==0.13.4
vine==5.0.0
wcwidth==0.2.5